Fix problems with growFrame assuming screen origins at 0,0
This commit fixes some issues with how growFrame sets bounds for the various
frame parameters assuming that the top left corner of the screen is 0,0. This
affects growing frames on secondary monitors in certain positions. The fix works
by making sure that screen frame coordinates are properly inserted into
calculations.